Version 0.1.4 - Fixed many bugs on BaumWelch class
BaumWelch can finally output a model, but I'm not sure if the created
model is the wanted model (still have to do more tests). I also need to
rewrite the execute method in this class, since I've to manually execute
a step and process all the results (expecially alpha) for other
purposes.