Version 0.1.8
* Added logging options for Gamma and Psi formulas. It's a quite
extensive log if lots of observations and states are used, so it's
disabled for default. I'll rework the debug options in order to take
account of this fact.
* Added a partial logging option for Baum-Welch
* Removed date and time from the log and now everything is saved to file
without being displayed on the console (thanks to Jarrod Roberson @
Stack Overflow for the tip)
* Alpha formula is under review, since I'm discussing with my colleagues
what formula I should use between 102 or 103 on Rabiner's paper
* Added a Rescaler for pre-processing certain types of data for an
example I'll add in future